Conversions.

+++ branches/trunk/GUI/spot-on-f.cc
@@ -1663,7 +1663,7 @@ void spoton::slotPassphraseChanged(const QString &text)
   spoton_pacify pacify(text.toStdString());
 
   result = 100.00 * pacify.evaluate() / maximum;
-  m_ui.passphrase_strength_indicator->setValue(result);
+  m_ui.passphrase_strength_indicator->setValue(static_cast<int> (result));
 
   if(result >= 0.00 && result <= 25.00)
     m_ui.passphrase_strength_indicator->setStyleSheet
